Fixes #20, Fixes #19, Fixes #18, Fixes #7, and accidentally, due to game speed retweaking, Fixes #6. Bonzo drops a lot faster, and collapsibles also collapse at a better speed. All in all, this update makes things look AND perform closer to the original game.

+++ Monkey Shines/src/org/erikaredmark/monkeyshines/Sprite.java
@@ -30,8 +30,8 @@ public class Sprite {
 	private       Point2D currentLocation;
 	
 	// These values are SET as the final speed (sprite speed times game constant multiplier)
-	private       double speedX;
-	private       double speedY;
+	private int speedX;
+	private int speedY;
 	
 	private final AnimationType animationType;
 	private final AnimationSpeed animationSpeed;
